@@ -11,119 +11,31 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
 		<!-- maven -->
-		<maven.compiler.source>11</maven.compiler.source>
-		<maven.compiler.target>11</maven.compiler.target>
-		<maven.compiler.version>3.8.0</maven.compiler.version>
-		<!-- <maven.surefire.plugin.version>2.22.0</maven.surefire.plugin.version> -->
-
-		<maven.jar.plugin.version>3.0.2</maven.jar.plugin.version>
-		<maven.war.plugin.version>3.1.0</maven.war.plugin.version>
-		<!-- <maven.jacoco.plugin.version>0.8.1</maven.jacoco.plugin.version> -->
-		<!-- <maven.sonar.plugin.version>3.7.0.1746</maven.sonar.plugin.version> -->
-		<maven.javadoc.version>3.2.0</maven.javadoc.version>
-		<maven-shade-plugin.version>2.3</maven-shade-plugin.version>
-
-		<!-- spring -->
-		<spring.boot.version>2.0.2.RELEASE</spring.boot.version>
-		<spring.data.jpa.version>2.0.7.RELEASE</spring.data.jpa.version>
-		<spring.security.test.version>5.0.5.RELEASE</spring.security.test.version>
-		<spring-cloud-config.version>2.0.4.RELEASE</spring-cloud-config.version>
-
-		<!-- swagger -->
-		<swagger.core.version>2.0.7</swagger.core.version>
-		<swagger.annotations.version>1.5.21</swagger.annotations.version>
-		<springfox.version>2.9.2</springfox.version>
-
-		<!-- vertx -->
-		<vertx.version>3.6.2</vertx.version>
-		<vertx.health.check.version>3.7.0</vertx.health.check.version>
-
-		<!-- javax -->
-		<javax.interceptor.version>1.2</javax.interceptor.version>
-		<javax.el-api.version>3.0.0</javax.el-api.version>
-		<javax.transaction.api.version>1.3</javax.transaction.api.version>
-		<javax.persistence.api.version>2.2</javax.persistence.api.version>
-		<javax.validation-api.version>2.0.1.Final</javax.validation-api.version>
-		<glassfish.javax.el>2.2.6</glassfish.javax.el>
-
-		<!-- data -->
-		<h2.version>1.4.197</h2.version>
-		<mysql.connector.version>5.1.46</mysql.connector.version>
-		<postgresql.version>42.2.2</postgresql.version>
-		<eclipselink.version>2.5.0</eclipselink.version>
-		<hibernate.version>5.2.17.Final</hibernate.version>
-		<hibernate.validator.version>6.0.12.Final</hibernate.validator.version>
-
-		<!-- test -->
-		<junit.version>4.12</junit.version>
-		<mockito.version>2.23.4</mockito.version>
-		<powermock.version>1.7.4</powermock.version>
-		<powermock.beta.version>2.0.7</powermock.beta.version>
-
-		<!-- logger -->
-		<logback.version>1.2.3</logback.version>
-		<slf4j.version>1.7.19</slf4j.version>
-		<log4j-over-slf4j>1.7.7</log4j-over-slf4j>
-		<jcl.to.slf4j.version>1.7.25</jcl.to.slf4j.version>
-
-		<!-- json -->
-		<jackson.version>2.9.5</jackson.version>
-		<jackson.datatype.version>2.9.8</jackson.datatype.version>
-		<jackson.databind>2.9.6</jackson.databind>
-		<json.utility.version>20180130</json.utility.version>
-		<json.schema.validator.version>2.2.10</json.schema.validator.version>
-		<json.version>20180813</json.version>
-		<jsonsimple.version>1.1.1</jsonsimple.version>
-
-		<!-- utilities -->
-		<commons-math3>3.6.1</commons-math3>
-		<commons-lang3>3.7</commons-lang3>
-		<commons-io>2.6</commons-io>
-		<commons-codec>1.11</commons-codec>
-		<commons-collections4.version>4.3</commons-collections4.version>
-		<commons.beanutils.version>1.9.2</commons.beanutils.version>
-		<commons.configuration2.version>2.2</commons.configuration2.version>
-		<apache.httpcomponents.version>4.5.6</apache.httpcomponents.version>
-		<guava.version>19.0</guava.version>
-		<lombok.version>1.18.32</lombok.version>
-		<jsch.version>0.1.54</jsch.version>
-		<sshd.version>1.4.0</sshd.version>
-		<itextcore.version>7.1.0</itextcore.version>
-		<itexthtml2pdf.version>2.0.0</itexthtml2pdf.version>
-		<itext.version>5.5.13</itext.version>
-		<freemarker.version>2.3.23</freemarker.version>
-		<velocity.version>1.7</velocity.version>
-		<velocity.tool.version>2.0</velocity.tool.version>
-		<datamapper.orika>1.5.2</datamapper.orika>
-		<modelmapper.version>2.1.1</modelmapper.version>
-		<bouncycastle.version>1.66</bouncycastle.version>
-		<icu4j.version>63.1</icu4j.version>
-		<seventeam-otp.version>1.0.0</seventeam-otp.version>
-		<google.zxing.version>3.3.3</google.zxing.version>
-		<hadoop-client>3.1.0</hadoop-client>
-		<htrace-core4>4.1.0-incubating</htrace-core4>
-		<aws.javasdk.version>1.11.368</aws.javasdk.version>
-		<s3mock.version>0.2.4</s3mock.version>
-		<ignite.version>2.3.0</ignite.version>
-		<findbugs.version>3.0.1</findbugs.version>
-		<byte.buddy.version>1.9.12</byte.buddy.version>
-		<io.jsonwebtoken.jjwt.version>0.6.0</io.jsonwebtoken.jjwt.version>
-		<apache.directory.api.version>2.0.0.AM2</apache.directory.api.version>
-		<aspectjweaver.version>1.8.12</aspectjweaver.version>
-		<micrometer.core.version>1.4.2</micrometer.core.version>
-		<micrometer.registry.prometheus.version>1.4.2</micrometer.registry.prometheus.version>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<kernel.core.version>1.2.0.1</kernel.core.version>
-		<kernel.idgenerator.vid.version>1.2.0.1</kernel.idgenerator.vid.version>
+		<maven.compiler.source>21</maven.compiler.source>
+		<maven.compiler.target>21</maven.compiler.target>
+		<kernel.core.version>1.2.1-SNAPSHOT</kernel.core.version>
+		<kernel.idgenerator.vid.version>1.2.1-SNAPSHOT</kernel.idgenerator.vid.version>
 		<ceylon.complete.version>1.3.2</ceylon.complete.version>
 		<vertx.lang.ceylon.version>3.4.1</vertx.lang.ceylon.version>
-		<kernel.templatemanager.velocity.version>1.2.0.1</kernel.templatemanager.velocity.version>
-		<kernel.crypto.signature.version>1.2.0.1</kernel.crypto.signature.version>
+		<kernel.templatemanager.velocity.version>1.2.1-SNAPSHOT</kernel.templatemanager.velocity.version>
+		<kernel.crypto.signature.version>1.2.1-SNAPSHOT</kernel.crypto.signature.version>
 		<modelmapper.version>2.3.7</modelmapper.version>
-		<kernel.logger.version>1.2.0.1</kernel.logger.version>
-		<jacoco.maven.plugin.version>0.8.5</jacoco.maven.plugin.version>
+		<kernel.logger.version>1.2.1-SNAPSHOT</kernel.logger.version>
+		<jacoco.maven.plugin.version>0.8.11</jacoco.maven.plugin.version>
+		<prometheus.version>1.4.2</prometheus.version>
 		<sonar.coverage.exclusions>**/constant/**,**/config/**,**/httpfilter/**,**/cache/**,**/entity/**,**/model/**,**/exception/**,**/repository/**,**/verticle/**,**/spi/**,"**/proxy/**","**/entities/**","**/filter/**","**/util/**","**/verifier/**","**/IDGeneratorVertxApplication.java"</sonar.coverage.exclusions>
 	</properties>
+     	<dependencyManagement>
+		<dependencies>
+			<dependency>
+				<groupId>io.mosip.kernel</groupId>
+				<artifactId>kernel-bom</artifactId>
+				<version>1.2.1-java21-SNAPSHOT</version>
+				<type>pom</type>
+			    <scope>import</scope>
+			</dependency>
+		</dependencies>
+	</dependencyManagement>
 	<artifactId>kernel-idgenerator-service</artifactId>
 	<version>1.2.1-SNAPSHOT</version>
 	<dependencies>
