@@ -3,7 +3,7 @@ package io.mosip.kernel.idgenerator.test.config;
 import java.util.HashMap;
 import java.util.Map;
 
-import javax.persistence.EntityManagerFactory;
+import jakarta.persistence.EntityManagerFactory;
 import javax.sql.DataSource;
 
 import org.mockito.Mockito;
